# docs/api/swagger.yaml
basePath: /
definitions:
...
    InstanceConfigurationEmojis:
        properties:
            emoji_size_limit:
                description: Max allowed emoji image size in bytes.
                example: 51200
                format: int64
                type: integer
                x-go-name: EmojiSizeLimit
        title: InstanceConfigurationEmojis models instance emoji config parameters.
        type: object
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
    adminEmoji:
        properties:
            category:
                description: Used for sorting custom emoji in the picker.
                example: blobcats
                type: string
                x-go-name: Category
            content_type:
                description: The MIME content type of the emoji.
                example: image/png
                type: string
                x-go-name: ContentType
            disabled:
                description: True if this emoji has been disabled by an admin action.
                example: false
                type: boolean
                x-go-name: Disabled
            domain:
                description: The domain from which the emoji originated. Only defined for remote domains, otherwise key will not be set.
                example: example.org
                type: string
                x-go-name: Domain
            id:
                description: The ID of the emoji.
                example: 01GEM7SFDZ7GZNRXFVZ3X4E4N1
                type: string
                x-go-name: ID
            shortcode:
                description: The name of the custom emoji.
                example: blobcat_uwu
                type: string
                x-go-name: Shortcode
            static_url:
                description: A link to a static copy of the custom emoji.
                example: https://example.org/fileserver/emojis/blogcat_uwu.png
                type: string
                x-go-name: StaticURL
            total_file_size:
                description: The total file size taken up by the emoji in bytes, including static and animated versions.
                example: 69420
                format: int64
                type: integer
                x-go-name: TotalFileSize
            updated_at:
                description: Time when the emoji image was last updated.
                example: "2022-10-05T09:21:26.419Z"
                type: string
                x-go-name: UpdatedAt
            uri:
                description: The ActivityPub URI of the emoji.
                example: https://example.org/emojis/016T5Q3SQKBT337DAKVSKNXXW1
                type: string
                x-go-name: URI
            url:
                description: Web URL of the custom emoji.
                example: https://example.org/fileserver/emojis/blogcat_uwu.gif
                type: string
                x-go-name: URL
            visible_in_picker:
                description: Emoji is visible in the emoji picker of the instance.
                example: true
                type: boolean
                x-go-name: VisibleInPicker
        title: AdminEmoji models the admin view of a custom emoji.
        type: object
        x-go-name: AdminEmoji
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
    instanceConfigurationAccounts:
        properties:
            allow_custom_css:
                description: Whether or not accounts on this instance are allowed to upload custom CSS for profiles and statuses.
                example: false
                type: boolean
                x-go-name: AllowCustomCSS
            max_featured_tags:
                description: |-
                    The maximum number of featured tags allowed for each account.
                    Currently not implemented, so this is hardcoded to 10.
                format: int64
                type: integer
                x-go-name: MaxFeaturedTags
        title: InstanceConfigurationAccounts models instance account config parameters.
        type: object
        x-go-name: InstanceConfigurationAccounts
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
    instanceV1:
        properties:
            account_domain:
                description: |-
                    The domain of accounts on this instance.
                    This will not necessarily be the same as
                    simply the Host part of the URI.
                example: example.org
                type: string
                x-go-name: AccountDomain
            approval_required:
                description: New account registrations require admin approval.
                type: boolean
                x-go-name: ApprovalRequired
            configuration:
                $ref: '#/definitions/instanceV1Configuration'
            contact_account:
                $ref: '#/definitions/account'
            description:
                description: |-
                    Description of the instance.

                    Should be HTML formatted, but might be plaintext.

                    This should be displayed on the 'about' page for an instance.
                type: string
                x-go-name: Description
            email:
                description: An email address that may be used for inquiries.
                example: admin@example.org
                type: string
                x-go-name: Email
            invites_enabled:
                description: Invites are enabled on this instance.
                type: boolean
                x-go-name: InvitesEnabled
            languages:
                description: Primary language of the instance.
                example: en
                items:
                    type: string
                type: array
                x-go-name: Languages
            max_toot_chars:
                description: |-
                    Maximum allowed length of a post on this instance, in characters.

                    This is provided for compatibility with Tusky and other apps.
                example: 5000
                format: uint64
                type: integer
                x-go-name: MaxTootChars
            registrations:
                description: New account registrations are enabled on this instance.
                type: boolean
                x-go-name: Registrations
            short_description:
                description: |-
                    A shorter description of the instance.

                    Should be HTML formatted, but might be plaintext.

                    This should be displayed on the instance splash/landing page.
                type: string
                x-go-name: ShortDescription
            stats:
                additionalProperties:
                    format: int64
                    type: integer
                description: 'Statistics about the instance: number of posts, accounts, etc.'
                type: object
                x-go-name: Stats
            thumbnail:
                description: URL of the instance avatar/banner image.
                example: https://example.org/files/instance/thumbnail.jpeg
                type: string
                x-go-name: Thumbnail
            thumbnail_description:
                description: Description of the instance thumbnail.
                example: picture of a cute lil' friendly sloth
                type: string
                x-go-name: ThumbnailDescription
            thumbnail_type:
                description: MIME type of the instance thumbnail.
                example: image/png
                type: string
                x-go-name: ThumbnailType
            title:
                description: The title of the instance.
                example: GoToSocial Example Instance
                type: string
                x-go-name: Title
            uri:
                description: The URI of the instance.
                example: https://gts.example.org
                type: string
                x-go-name: URI
            urls:
                $ref: '#/definitions/instanceV1URLs'
            version:
                description: |-
                    The version of GoToSocial installed on the instance.

                    This will contain at least a semantic version number.

                    It may also contain, after a space, the short git commit ID of the running software.
                example: 0.1.1 cb85f65
                type: string
                x-go-name: Version
        title: InstanceV1 models information about this instance.
        type: object
        x-go-name: InstanceV1
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
    instanceV1Configuration:
        properties:
            accounts:
                $ref: '#/definitions/instanceConfigurationAccounts'
            emojis:
                $ref: '#/definitions/InstanceConfigurationEmojis'
            media_attachments:
                $ref: '#/definitions/instanceConfigurationMediaAttachments'
            polls:
                $ref: '#/definitions/instanceConfigurationPolls'
            statuses:
                $ref: '#/definitions/instanceConfigurationStatuses'
        title: InstanceV1Configuration models instance configuration parameters.
        type: object
        x-go-name: InstanceV1Configuration
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
    instanceV1URLs:
        properties:
            streaming_api:
                description: Websockets address for status and notification streaming.
                example: wss://example.org
                type: string
                x-go-name: StreamingAPI
        title: InstanceV1URLs models instance-relevant URLs for client application consumption.
        type: object
        x-go-name: InstanceV1URLs
        x-go-package: github.com/superseriousbusiness/gotosocial/internal/api/model
...
host: example.org
info:
    contact:
        email: admin@gotosocial.org
        name: GoToSocial Authors
    license:
        name: AGPL3
        url: https://www.gnu.org/licenses/agpl-3.0.en.html
    title: GoToSocial Swagger documentation.
    version: REPLACE_ME
paths:
...
    /api/v1/admin/custom_emojis/{id}:
        delete:
            description: |-
                Emoji with the given ID will no longer be available to use on the instance.

                If you just want to update the emoji image instead, use the `/api/v1/admin/custom_emojis/{id}` PATCH route.

                To disable emojis from **remote** instances, use the `/api/v1/admin/custom_emojis/{id}` PATCH route.
            operationId: emojiDelete
            parameters:
                - description: The id of the emoji.
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: The deleted emoji will be returned to the caller in case further processing is necessary.
                    schema:
                        $ref: '#/definitions/adminEmoji'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - admin
            summary: Delete a **local** emoji with the given ID from the instance.
            tags:
                - admin
        get:
            operationId: emojiGet
            parameters:
                - description: The id of the emoji.
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: A single emoji.
                    schema:
                        $ref: '#/definitions/adminEmoji'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            summary: Get the admin view of a single emoji.
            tags:
                - admin
        patch:
            consumes:
                - multipart/form-data
            description: |-
                Action performed depends upon the action `type` provided.

                `disable`: disable a REMOTE emoji from being used/displayed on this instance. Does not work for local emojis.

                `copy`: copy a REMOTE emoji to this instance. When doing this action, a shortcode MUST be provided, and it must
                be unique among emojis already present on this instance. A category MAY be provided, and the copied emoji will then
                be put into the provided category.

                `modify`: modify a LOCAL emoji. You can provide a new image for the emoji and/or update the category.

                Local emojis cannot be deleted using this endpoint. To delete a local emoji, check DELETE /api/v1/admin/custom_emojis/{id} instead.
            operationId: emojiUpdate
            parameters:
                - description: The id of the emoji.
                  in: path
                  name: id
                  required: true
                  type: string
                - description: |-
                    Type of action to be taken. One of: (`disable`, `copy`, `modify`).
                    For REMOTE emojis, `copy` or `disable` are supported.
                    For LOCAL emojis, only `modify` is supported.
                  in: formData
                  name: type
                  required: true
                  type: string
                - description: The code to use for the emoji, which will be used by instance denizens to select it. This must be unique on the instance. Works for the `copy` action type only.
                  in: formData
                  name: shortcode
                  pattern: \w{2,30}
                  type: string
                - description: A new png or gif image to use for the emoji. Animated pngs work too! To ensure compatibility with other fedi implementations, emoji size limit is 50kb by default. Works for LOCAL emojis only.
                  in: formData
                  name: image
                  type: file
                - description: Category in which to place the emoji. 64 characters or less. If a category with the given name doesn't exist yet, it will be created.
                  in: formData
                  name: category
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: The updated emoji.
                    schema:
                        $ref: '#/definitions/adminEmoji'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - admin
            summary: Perform admin action on a local or remote emoji known to this instance.
            tags:
                - admin
...
    /api/v1/instance:
        get:
            operationId: instanceGetV1
            produces:
                - application/json
            responses:
                "200":
                    description: Instance information.
                    schema:
                        $ref: '#/definitions/instanceV1'
                "406":
                    description: not acceptable
                "500":
                    description: internal error
            summary: View instance information.
            tags:
                - instance
        patch:
            consumes:
                - multipart/form-data
            description: This requires admin permissions on the instance.
            operationId: instanceUpdate
            parameters:
                - allowEmptyValue: true
                  description: Title to use for the instance.
                  in: formData
                  maximum: 40
                  name: title
                  type: string
                - allowEmptyValue: true
                  description: Username of the contact account. This must be the username of an instance admin.
                  in: formData
                  name: contact_username
                  type: string
                - allowEmptyValue: true
                  description: Email address to use as the instance contact.
                  in: formData
                  name: contact_email
                  type: string
                - allowEmptyValue: true
                  description: Short description of the instance.
                  in: formData
                  maximum: 500
                  name: short_description
                  type: string
                - allowEmptyValue: true
                  description: Longer description of the instance.
                  in: formData
                  maximum: 5000
                  name: description
                  type: string
                - allowEmptyValue: true
                  description: Terms and conditions of the instance.
                  in: formData
                  maximum: 5000
                  name: terms
                  type: string
                - description: Thumbnail image to use for the instance.
                  in: formData
                  name: thumbnail
                  type: file
                - description: Image description of the submitted instance thumbnail.
                  in: formData
                  name: thumbnail_description
                  type: string
                - description: Header image to use for the instance.
                  in: formData
                  name: header
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: The newly updated instance.
                    schema:
                        $ref: '#/definitions/instance'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - admin
            summary: Update your instance information and/or upload a new avatar/header for the instance.
            tags:
                - instance
...
    /api/v1/statuses:
        post:
            consumes:
                - application/json
                - application/xml
                - application/x-www-form-urlencoded
            description: |-
                The parameters can also be given in the body of the request, as JSON, if the content-type is set to 'application/json'.
                The parameters can also be given in the body of the request, as XML, if the content-type is set to 'application/xml'.
            operationId: statusCreate
            parameters:
                - description: |-
                    Text content of the status.
                    If media_ids is provided, this becomes optional.
                    Attaching a poll is optional while status is provided.
                  in: formData
                  name: status
                  type: string
                  x-go-name: Status
                - description: |-
                    Array of Attachment ids to be attached as media.
                    If provided, status becomes optional, and poll cannot be used.

                    If the status is being submitted as a form, the key is 'media_ids[]',
                    but if it's json or xml, the key is 'media_ids'.
                  in: formData
                  items:
                    type: string
                  name: media_ids
                  type: array
                  x-go-name: MediaIDs
                - description: ID of the status being replied to, if status is a reply.
                  in: formData
                  name: in_reply_to_id
                  type: string
                  x-go-name: InReplyToID
                - description: Status and attached media should be marked as sensitive.
                  in: formData
                  name: sensitive
                  type: boolean
                  x-go-name: Sensitive
                - description: |-
                    Text to be shown as a warning or subject before the actual content.
                    Statuses are generally collapsed behind this field.
                  in: formData
                  name: spoiler_text
                  type: string
                  x-go-name: SpoilerText
                - description: Visibility of the posted status.
                  in: formData
                  name: visibility
                  type: string
                  x-go-name: Visibility
                - description: |-
                    ISO 8601 Datetime at which to schedule a status.
                    Providing this parameter will cause ScheduledStatus to be returned instead of Status.
                    Must be at least 5 minutes in the future.
                  in: formData
                  name: scheduled_at
                  type: string
                  x-go-name: ScheduledAt
                - description: ISO 639 language code for this status.
                  in: formData
                  name: language
                  type: string
                  x-go-name: Language
                - description: Format to use when parsing this status.
                  in: formData
                  name: format
                  type: string
                  x-go-name: Format
                - description: This status will be federated beyond the local timeline(s).
                  in: query
                  name: federated
                  type: boolean
                  x-go-name: Federated
                - description: This status can be boosted/reblogged.
                  in: query
                  name: boostable
                  type: boolean
                  x-go-name: Boostable
                - description: This status can be replied to.
                  in: query
                  name: replyable
                  type: boolean
                  x-go-name: Replyable
                - description: This status can be liked/faved.
                  in: query
                  name: likeable
                  type: boolean
                  x-go-name: Likeable
            produces:
                - application/json
            responses:
                "200":
                    description: The newly created status.
                    schema:
                        $ref: '#/definitions/status'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - write:statuses
            summary: Create a new status.
            tags:
                - statuses
    /api/v1/statuses/{id}:
        delete:
            description: |-
                The deleted status will be returned in the response. The `text` field will contain the original text of the status as it was submitted.
                This is useful when doing a 'delete and redraft' type operation.
            operationId: statusDelete
            parameters:
                - description: Target status ID.
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: The status that was just deleted.
                    schema:
                        $ref: '#/definitions/status'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - write:statuses
            summary: Delete status with the given ID. The status must belong to you.
            tags:
                - statuses
        get:
            operationId: statusGet
            parameters:
                - description: Target status ID.
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: The requested status.
                    schema:
                        $ref: '#/definitions/status'
                "400":
                    description: bad request
                "401":
                    description: unauthorized
                "403":
                    description: forbidden
                "404":
                    description: not found
                "406":
                    description: not acceptable
                "500":
                    description: internal server error
            security:
                - OAuth2 Bearer:
                    - read:statuses
            summary: View status with the given ID.
            tags:
                - statuses
...
schemes:
    - https
    - http
securityDefinitions:
    OAuth2 Application:
        flow: application
        scopes:
            write:accounts: grants write access to accounts
        tokenUrl: https://example.org/oauth/token
        type: oauth2
    OAuth2 Bearer:
        authorizationUrl: https://example.org/oauth/authorize
        flow: accessCode
        scopes:
            admin: grants admin access to everything
            admin:accounts: grants admin access to accounts
            read: grants read access to everything
            read:accounts: grants read access to accounts
            read:blocks: grant read access to blocks
            read:custom_emojis: grant read access to custom_emojis
            read:favourites: grant read access to favourites
            read:follows: grant read access to follows
            read:media: grant read access to media
            read:notifications: grants read access to notifications
            read:search: grant read access to searches
            read:statuses: grants read access to statuses
            read:streaming: grants read access to streaming api
            read:user: grants read access to user-level info
            write: grants write access to everything
            write:accounts: grants write access to accounts
            write:blocks: grants write access to blocks
            write:follows: grants write access to follows
            write:media: grants write access to media
            write:statuses: grants write access to statuses
            write:user: grants write access to user-level info
        tokenUrl: https://example.org/oauth/token
        type: oauth2
swagger: "2.0"